ぐちゃぐちゃになりそうなのでカスタムフック化
