+++ client/src/pages/view/view.module.css
@@ -72,7 +72,6 @@
   position: relative;
   z-index: 1;
   max-width: 100%;
-  border-radius: 8px;
   transition: 0.3s;
 }
